feat: rabc permission sync

+++ user-service/src/main/java/com/aitschool/user/Rabc/model/Permission.java
@@ -2,12 +2,10 @@ package com.aitschool.user.Rabc.model;
 
 import com.fasterxml.jackson.annotation.JsonBackReference;
 import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import jakarta.persistence.*;
 import lombok.Data;
 
 import java.util.Date;
-import java.util.List;
 
 @Entity
 @Data
@@ -26,7 +24,7 @@ public class Permission {
 
     private String description;
 
-    private Long sort;
+    private Integer sort;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "group_id", nullable = false)
